Print help/version without error prefix (#455)
* xargs: print help/version without error prefix

xargs, when executed with -h/--help or -V/--version, is printing an
"Error: " prefix before the help and version text. The exit code is
also incorrect (1, instead of 0).

Display the help/version text without the erroneous prefix, and exit
with the correct exit code (success, 0).

Add tests for this bug to prevent regressions.