matches/exec: change tests and check_path implementation
renamed: check_path_integrity -> check_path_entries_absolute

The function check_path_entries_absolute now only looks for paths that are not absolute or that are empty.
In addition, if possible, if there is any error related to the PATH, show the exact segment where the error occurred.

tests:

The tests that used to reside in exec_unit_tests.rs are now in exec.rs, directly handling a fake of the PATH for testing,
eliminating the need to mark tests with #[serial] or use unsafe { env::set_var(...) }.

Co-authored-by: Tavian <tavianator@tavianator.com>